requirements.txt
attrs==21.2.0
beaker-pyteal==0.5.2
blinker==1.4
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docstring-parser==0.14.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
msgpack==1.0.4
netifaces==0.11.0
oauthlib==3.2.0
py-algorand-sdk==2.0.0
pycparser==2.21
pycryptodomex==3.17
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyteal==0.22.0
python-apt==2.4.0
python-dotenv==0.19.2
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
semantic-version==2.10.0
six==1.16.0
systemd-python==234
texttable==1.6.4
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websocket-client==1.2.3
zipp==1.0.0
